node/manager: Extract node checkpoint implementation
Move the node checkpoint implementation from manager.go into a dedicated
file without changing behavior. This prepares checkpoint ownership to move
to the Linux datapath in a separately reviewable patch.

AIL:3
Related: https://github.com/cilium/cilium/issues/41744
